pyproject.toml
[tool.poetry]
name = "creme-server"
version = "0.1.0"
description = ""
authors = ["rigyeonghong <zxcvbee@naver.com>"]

[tool.poetry.dependencies]
python = "3.8"
Django = "^4.0.2"
djangorestframework = "^3.13.1"
cmake = "^3.22.2"
dlib = "19.8.2"
tensorflow = "^2.8.0"
imutils = "^0.5.4"
pandas = "^1.4.1"
opencv-python = "^4.5.5"
requests = "^2.27.1"
rmn = "^3.0.3"
torch = "^1.10.2"
Pillow = "^9.0.1"
nanoid = "^2.0.0"
drf-yasg = "^1.20.0"
django-cors-headers = "^3.11.0"

[tool.poetry.dev-dependencies]
flake8 = "4.0.1"
isort = "5.10.1"
black = "21.12b0"
django-silk = "4.2.0"
pre-commit = "2.17.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

# isort
[tool.isort]
include_trailing_comma = true
line_length = 88
profile = "black"
skip_glob = "**/migrations/*.py"
use_parentheses = false

# black
[tool.black]
extend-exclude = '''
     migrations
'''
include = '\.pyi?$'
line-length = 88
